BackEnd Sept 20 update
fixed some bugs and did some simple tests.
Now there are only two incomplete places:
1. I cannot upload my hash table implementation now because of that assignment. Therefore I am using HashMap right now
2. I need some usernames and passwords that I can load them in the constructor. So, Timothy please add a function that can load account information

+++ src/BackEnd/BackEnd.java
@@ -63,7 +63,7 @@ public interface BackEnd {
     return the subtotal of the cart
     it should be positive, but it will be -1 if there is unexpected error
     */
-    public int subtotal();
+    public float subtotal();
 
     /*
     Clear the cart and calculate new quantities of items in the cart
